trying to reduce precision warnings about conversion from double to float
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@2998 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/air/air.h
@@ -465,7 +465,7 @@ AIR_EXPORT double airTime();
 AIR_EXPORT double airCbrt(double);
 AIR_EXPORT double airSgnPow(double, double);
 AIR_EXPORT int airSgn(double);
-AIR_EXPORT int airLog2(float n);
+AIR_EXPORT int airLog2(double n);
 AIR_EXPORT void airBinaryPrintUInt(FILE *file, int digits, unsigned int N);
 AIR_EXPORT double airErf(double x);
 AIR_EXPORT double airGaussian(double x, double mean, double stdv);
